Fixed a bunch of bugs in localify.  We call GC-GROVEL-STACK before doing localify so that GGS doesn't get confused by objects being trashed with GC forwards.
In TRANSPORT-FUNCTION, we must be prepared to find closure-entry objects (and
treat them the same as normal entries.)  Also, we must check if the entry for a
closure is pure before transporting the entry.

In various places, I had to flush uses of INFO, since this accesses global
variables (*info-environment*).