Picked up 16x32 divide stuff from Bill (but it is conditionalize out).  Also other misc changes from Bill.
Added BIGNUM-REPLACE and made people use it.  Add WITH-BIGNUM-BUFFERS and made
people use it so that they would be reentrant.  Fixed MAKE-GCD-BIGNUM-ODD to
not assume that there is a 1 in the first digit.  Added misc declarations.
Changed shifting stuff to use the new %DIGIT-LOGICAL-SHIFT-RIGHT.
Un-commented-out GCD and multi-digit bignum truncate.