Generalized CODE-PACKAGE-BREAKDOWN to also do breakdowns by file, and changed its name to CODE-BREAKDOWN.
Added a bunch of functions for finding the indegree and outdegree of nodes in
the call graph (i.e. how many times functions are called, and how many calls
functions make.)  Also added some stuff that uses Scott's histogram package (in
the library) this is #+NIL so that it will compile without.  See
MEMORY-HISTOGRAM which can print various histograms of object statistics.