Changed DEFSETF and DEFINE-SETF-METHOD to call a common subfunction, %DEFINE-SETF-MACRO, which checks for any existing definition as a slot accessor or other setf function.
Changed GET-SETF-METHOD to ignore setf macros (always global) when there is a
local macro or function.

Changed SETF to always call GET-SETF-METHOD, instead of special-casing setf
inverses.  I was tired to keeping the duplicated code consistent.

Generalized the special casing in SETF of APPLY.  We now go for the SETF
function when called on any symbol in the Lisp package.  Now SETF of APPLY will
work on all Common Lisp functions, instead of just two.