When doing the labels hack in converting global inlines, respect :NOTINLINE declarations.
Changed declaration processing to treat a FUNCTION declaration as an ordinary
variably type declaration (as proposed by X3J13).  The old semantics is still
obtained when the second arg to the declaration is a list (as it always would
be in the old usage.)

Added support for the CONSTANT-FUNCTION declaration.