Removed declaration for non-existent variable focus-p in HUNK-MOUSE-ENTERED and HUNK-MOUSE-LEFT.
Added proportion argument to MAKE-BITMAP-WINDOW,
MAYBE-MAKE-X-WINDOW-AND-PARENT, and CREATE-WINDOW-FROM-CURRENT.

CREATE-WINDOW-FROM-CURRENT now creates a window according to its new proportion
argument instead of simply splitting the current window in two.  It returns nil
without doing anything if either window is too small.

WINDOW-GROUP-CHANGED no longer unifies the sizes of window when the user
resizes a group.  It now tries to distribute the new size of the group
according to the proportions of the old size consumed by the windows.