Fixed lossage concerning the following compiler output:    Definition has N args, but the previous definition had M. This used to be a note, but it should be a warning.
If it is a warning that users change the definitions of structures, then it
should be a warning when users change the definitions of functions.  In both
cases, code might be lying around assuming an incorrect interface.  I think
this is equally dangerous.  Also, changing the redefinition of arg counts to a
warning makes it more compatible with the warning that someone supposedly
called a routine with the wrong number of args.