When optimizing combinations, don't do the "efficiency hack" of checking whether any args have reoptimize set.  Most calls to IR1-OPTIMIZE-COMBINATION that this inhibits are actually desirable.
Deleted assertion in JOIN-SUCCESSOR-IF-POSSIBLE.  This really can happen, and
DELETE-CONTINUATION should do the right thing.

Changed MAYBE-TERMINATE-BLOCK to take an explicit IR1-P flag.  Fixed it not to
consider deleted continuations to have type NIL.  Use DELETE-CONTINUATION on
the old continuation if it has no references and we aren't in IR1 conversion.