Changed NN-REPLY-CLEANUP-SPLIT-WINDOWS to delete the message-window instead of the reply-window.
Split CONNECT-TO-NNTP into two: RAW-CONNECT-TO-NNTP and CONNECT-TO-NNTP.  The
first is the same as the original with the addition of specifying a timeout
value to MAKE-FD-STREAM.  The second binds a handler for the IO-TIMEOUT
condition and gives the user an error message.

Added Hemlock variable "Netnews NNTP Timeout Period" to control how long
Netnews will wait (while connecting to NNTP) before timing out.

Changed *nntp-server* into Hemlock variable "Netnews NNTP Server".

Added command "Netnews Reply to Sender in Other Window".

Added functions NN-SETUP-FOR-REPLY-BY-MAIL and NN-REPLY-TO-SENDER to hold
common code for "Netnews Forward Message", "Netnews Reply to Sender", and
"Netnews Reply to Sender in Other Window".

Changed default bindings to use "Netnews Reply to Group in Other Window" and
"Netnews Reply to Sender in Other Window"