Changed COMMUTATIVE-ARG-SWAP to actually splice in the constant arg, so that variable substitution can't swap it back again.
Fixed multiple recoding to include many TRULY-THE's in the expansion so that
the resulting shift-and-add code would actually open coded.

Added comprehensive handling of arithmetic and logical identities when an arg
is -1, 0 or +1.

Fixed the (- 0 x) ==> (%negate x) transform to be inhibited when the 0 is more
contagious than X, or is -0.0.