Added a doc string for LOAD.  Added ANSI features *LOAD-TRUENAME*, *LOAD-PATHNAME* and *LOAD-PRINT*.  As per ANSI, bind *READTABLE* to itself to make assignments file-local.
Added new variables EXT:*SOURCE-FILE-TYPES* and EXT:*OBJECT-FILE-TYPES*.  When
no file type is specified, LOAD tries the types in these lists to locate the
source and object files.  LOAD now recognizes source types "l", "cl" and "lsp"
in addition to "lisp".

We now bind *LOAD-PRINT*, *LOAD-VERBOSE* and EXT:*LOAD-IF-SOURCE-NEWER*
conditionally depending (via PROGV) on whether they corresponding argument is
actually specified.  This allows these variables to be altered by load (such as
of an init file.)

The compiler OPTIMIZE policy is now bound during load, so proclamations in a
file don't leave the global policy clobbered when the load is finished.

Changed the :IF-SOURCE-NEWER option to signal an error and use restarts, rather
than PROMPT-FOR-Y-OR-N.  Fixed the load source case to actually load the
source, rather than loading the object as a source file...