Some style cleanups.
Change the offset used by the sap VOPS from an unsigned-num to a
signed-num. This helps fixes some SAP functions for negative offsets
which are useful on the x86 with its stack growing down. The offset is
currently declared to be a index (unsigned-byte 29) in
compiler/saptran.lisp so this should work without change, but to use
the negative offsets requires complementary changes to saptran (index
-> fixnum).