Updates from Raymond Toy:
Elfun-derive-type-union didn't properly handle the case of
elfun-derive-type-1 returning a union type like (or float (complex
float)) - it does now.

Elfun-derive-type-1 is slightly enhanced so that if the input is a
complex type, the result is a complex type of the same.  Before, we
just returned default-type.  Also, it's a bit smarter now so that if
the result is supposed to be (or float (complex float)), we try to
refine float to single-float or double-float if possible.