Have the functions for unbinding and rebinding the bind stack check that the symbol slot is actually a symbol as this slot may be zero if a bind or unbind is interrupted - have it ignore these for now so that it as least proceeds rather than giving a SIGSEGV fault.
Re-write of make-stack-group and stack-group-resume taking care not to
mess with unwind-protect during a stack group switch.

No longer need to disable the GENCGC pointer filter when starting the
SIGALRM yield interrupt because the object allocation should now be
interrupt safe now.

Have the SIGALRM handler return quickly if the interrupt nesting is
too great.