Update from Raymond Toy:
Fixed small bug in elfun-derive-type-1: When a complex number is
passed in, we don't want to just copy that as the result.  We want the
same type of complex number but with no bounds included.

New function float-or-complex-type to compute the union type (or float
(complex float)) with the appropriate type of float and bounds, if
needed.

At the end compiler/float-tran.lisp, the derive-type optimizers for
sin, cos, and tan, are never enabled, because of a misspelling of
propagate-fun-type (was progagate-fun-type). Plus a few fixes for
these.

Added expt derive-type optimizer, should handle all cases.

Added derive-type optimizers for conjugate and cis.