From: Raymond Toy <toy@rtp.ericsson.se>
While hacking on my lapack package, I needed to know a bit more about
the internals of CMUCL.  Here is a small patch that updates the
description of the array object.  It appears that the documentation is
wrong, and there is an extra slot in the array header to hold
fill-pointer-p.

I've also updated some of the type codes in the documentation.  I'm
sure about the array and number stuff, but not sure about the others.