For interval-div, handle the case where the denominator is 0 and the numerator is positive.  The result should be still be positive, but interval-div was incorrectly returning an unbounded interval. From Raymond Toy.
For two-arg-derive-type, when both arguments are member types, convert
both to numeric types and call the deriver function which often
results a tighter, less conservative, type. For example the division
of two positive numbers that include zero now returns a positive
type. The older behavior of calling the function with the two member
values is maintained but is now conditional on the
:conservative-float-type feature; also ensure that the :invalid trap
is masked out when calling the function even under the
negative-zero-is-not-zero feature. From Raymond Toy.