o Branch with prediction instructions are defined.
o On the V9, all branches are replaced with the corresponding
  branch-with-prediciton versions.  To minimize the number of
  changes, I did this by keeping the CMUCL instruction name the
  same, but it actually uses the prediction version.  This means,
  however, that you can't use the deprecated branch instructions
  on the V9.

  The disassembler, however, uses the branch-with-prediction
  instruction mnemonics, and should still be able to disassemble
  the deprecated branch instructions (I think).

o Added the conditional move instructions: conditional move
  integer/FP depending on the integer or FP condition code and
  conditional move integer/FP depending on the value of an integer
  register.

o The FCMP instructions can place its result into one of four FP
  condition registers on the V9.

o Unified the extended FP instructions by making the printer print
  the right thing instead of having several separate formats and
  macros for them.

o Added a :print-name keyword to the define-f3-inst macro.

o The ld instruction is printed as lduw on the V9 because that's
  what it is and does.

o Marked the deprecated V9 instructions but left them in.

o For the T instruction, set the type of the target to be the
  right size for the V9.  (It's only 7 bits instead of 13 on the V9.)