o Added truncate vops for V8 and V9.  The generic versions in   assembly/sparc/arith.lisp won't be used anymore on V8 or V9.
o fast-ash now uses a conditional move instruction on the V9.
  Should help since it gets rid of a branch instruction, but I
  can't really tell.

o The fast-* for signed and unsigned multiplication appropriately
  conditionalized to work on V8 and V9, as appropriate.  (As
  discussed in my earlier patch for assembly/sparc/arith.lisp).

o bignum::%digit-0-or-plusp uses a conditional move on the V9

o Some miscellaneous patches to the VOP guards so that the ones
  that require sparc-64 support are only done on sparc-64
  platforms (of which there aren't any supported yet).

o Added a bignum-floor vop for V8.