Marco Antoniotti says:
PARSE-NAMESTRING is supposed to accept also a string for the optional
HOST argument.  Right now this is not the case.

The following patch fixes this and seems to do the right thing when
HOST is a defined logical pathname host.

This is better than the current situation (as a matter of fact it is
necessary for some things I am doing), but it is not quite ANSI.
E.g. HOST could also be :UNSPECIFIC, but the current code does not
handle that case.  I did not add code for that, because, contrary to
the LOGICAL HOST case, I believe that the case for :UNSPECIFIC is too
implementation dependent.