From Eric Marsden:
There is an incorrect gc_assert in lisp/gc.c in the function
from_space_p. The argument is not necessarily a descriptor; it may be
an untagged pointer, in which case the assertion is false. This causes
rebuild to fail on SPARC when GC assertions are enabled.

This observation is due to Daniel Barlow (and is fixed in SBCL).