Added a cross-referencing facility to the compiler.
The new XREF package exports function WHO-CALLS, WHO-REFERENCES, WHO-BINDS
and WHO-SETS. These return structures of type XREF-CONTEXT, which contain
the name of the referencing context, the file it was referenced from (or NIL
when compiling from the toplevel), and the source-position in that file.

Also change the names of lambda nodes in IR1 for macros and compiler-macros
to be lists of the form

   (:macro macro-name)
   (:compiler-macro macro-name)

instead of strings of the form "DEFMACRO MACRO-NAME".

Documentation has been added to the User's Manual.

This change shouldn't cause any bootstrapping problems when building.