Entomotomy Bug: condition-reporting-not-via-print-object
This change causes condition printing (generally) and condition reporting
(in particular) to go via print-object, once PCL is loaded, as required
by ANSI.  Note that the :report option of define-condition still goes
through the old non-print-object mechanism (which is allowed by ANSI),
but that this mechanism is now invoked via a default method of print-object
specialized on condition.  Hence user-defined methods on print-object
are now effective, instead of being bypassed through the old mechanism.