o Remove old version of gencgc allocation macro. o Fix problem where the desired size is a number that won't fit in the   13-bit signed offset of an instruction.  Fortunately, we had a temp   register we could use.
However, we cannot currently stack allocate something that large.  The
code needs to be rearranged.  We also need to fix the code for
non-gencgc version, but I'm not going to do that.

There are also other issues that might crop up.  The number stack or
control stack could be an issue if there are enough locals or args.