Change the behaviour of EXT:LOAD-FOREIGN when called with a single shared library: call SYS::LOAD-OBJECT-FILE directly, without invoking the linker as previously. This should make it possible to support the reloading of dependent shared libraries upon startup of a saved image.
The new behaviour requires a function FILE-SHARED-LIBRARY-P to be defined. I
have provided an implementation that works for ELF platforms. This has been
tested with Linux/x86 and Solaris/SPARC, but is also enabled by
conditionalization for :bsd (should include FreeBSD and OpenBSD).

Also removed pmax support. IRIX and HPPA platforms will need to provide a
definition for F-S-L-P.